requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
django-appconf==1.1.0
django-compressor==4.5.1
lxml==5.3.1
packaging==24.2
psycopg2-binary==2.9.10
python-docx==1.1.2
rcssmin==1.1.2
requests==2.32.3
rjsmin==1.2.2
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
gunicorn==23.0.0